changes made
R-pipeline has been changed to now allow any city to run. The only thing that needs to be done is to state the city name at the beginning of the code.  The regional_epinow contains code that will run the regional estimates. However, it is better to use r_pipeline for each city. The toy dataset for the regional_epinow is also being uploaded to the same branch as toy_df.